requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.28.2
beautifulsoup4==4.12.2
selenium==4.9.1
openai==0.27.8
httpx==0.24.1
python-dotenv==1.0.0